CM-252: Simplify JMS priority handling to max 4 cases
The DAQ only supports four JMS priorities which are further described by
the given constants. This has influence also on the local message
buffering which is used to send value update in bunches. This adjustment
is necessary as the initial design has slightly changed over the years
to improve the overall sending performance.